Added models folder and renamed router name
