+++ server/models/index.js
@@ -1,0 +1,1 @@
+module.exports.getSpots = require('./getSpots.js').getSpots;
