+++ server/routes/controllers/getSpots.js
@@ -1,26 +1,15 @@
-const client = require('../../../db/index.js').client;
+const models = require('../../models/index.js');
 
-const getSpots = (req, res) => {
-  let query = `SELECT spot_id, host_id, lat, long, address, type, price, photo_url, ST_AsText(geom) AS geom FROM spots;`
-  return client.query(query)
+const getSpotsRoute = (req, res) => {
+  return models.getSpots()
     .then((spots) => {
-      console.log('spots in server', spots.rows);
-      let requiredSpots = spots.rows.map((spot) => {
-        return {
-          spot_id: spot.spot_id,
-          address: spot.address,
-          location: {
-            lat: Number(spot.lat),
-            lng: Number(spot.long)
-          }
-        }
-      });
-      console.log('REQUIRED SPOTS', requiredSpots)
-      res.send(requiredSpots)
+      console.log('SPOTS CONTROLLER', spots)
+      res.send(spots)
     })
     .catch ((err) => {
-      console.log('ERRRO GETTING SPOTS FROM DB in server', err)
+      console.log('ERRRO GETTING SPOTS FROM DB IN CONTROLLER', err)
+      res.sendStatus(500);
     })
 };
 
-module.exports.getSpots = getSpots;
+module.exports.getSpotsRoute = getSpotsRoute;
