+++ server/routes/routes.js
@@ -1,9 +1,9 @@
 
 const router = require('express').Router();
 const profileRequests = require('./controllers/profileRequests');
-const getSpots = require('./controllers/getSpots').getSpots;
+const getSpotsRoute = require('./controllers/getSpots').getSpotsRoute;
 
 router.get('/profile', profileRequests.getProfile);
-router.get('/spots', getSpots);
+router.get('/spots', getSpotsRoute);
 
 module.exports = router;
